--- a/optlib/org.c
+++ b/optlib/org.c
@@ -10,6 +10,17 @@
 
 static void initializeOrgParser (const langType language)
 {
+	addLanguageOptscriptToHook (language, SCRIPT_HOOK_PRELUDE,
+		"{{    /kindTable [\n"
+		"        /part\n"
+		"        /chapter\n"
+		"        /section\n"
+		"        /subsection\n"
+		"        /subsubsection\n"
+		"        /paragraph\n"
+		"        /subparagraph\n"
+		"    ] def\n"
+		"}}");
 
 	addLanguageRegexTable (language, "toplevel");
 	addLanguageRegexTable (language, "srcblock");
@@ -21,26 +32,11 @@ static void initializeOrgParser (const langType language)
 	                               "^#\\+begin_src[ ]+([a-zA-Z0-9][-#+a-zA-Z0-9]*)",
 	                               "", "", "{tenter=srcblock}{_guest=\\1,0end,}", NULL);
 	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*\\*\\*\\*\\*\\*\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "G", "", NULL);
-	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*\\*\\*\\*\\*\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "P", "", NULL);
-	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*\\*\\*\\*\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "b", "", NULL);
-	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*\\*\\*\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "u", "", NULL);
-	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*\\*\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "s", "", NULL);
-	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "c", "", NULL);
-	addLanguageTagMultiTableRegex (language, "toplevel",
-	                               "^\\*[ \t]+([[:graph:][:blank:]]+)([\n])?",
-	                               "\\1", "p", "", NULL);
+	                               "^(\\*{1,7})[ \t]+([[:graph:][:blank:]]+)([\n])?",
+	                               "", "", ""
+		"{{\n"
+		"    \\2 kindTable \\1 length 1 sub get @2 _tag _commit pop\n"
+		"}}", NULL);
 	addLanguageTagMultiTableRegex (language, "toplevel",
 	                               "^<<([^>]+)>>",
 	                               "\\1", "d", "", NULL);
@@ -109,6 +105,7 @@ extern parserDefinition* OrgParser (void)
 	def->patterns      = patterns;
 	def->aliases       = aliases;
 	def->method        = METHOD_NOT_CRAFTED|METHOD_REGEX;
+	def->useCork       = CORK_QUEUE;
 	def->kindTable     = OrgKindTable;
 	def->kindCount     = ARRAY_SIZE(OrgKindTable);
 	def->initialize    = initializeOrgParser;