--- a/optlib/org.ctags
+++ b/optlib/org.ctags
@@ -36,6 +36,19 @@
 # definition of levels
 --_tabledef-Org=toplevel
 --_tabledef-Org=srcblock
+
+--_prelude-Org={{
+    /kindTable [
+        /part
+        /chapter
+        /section
+        /subsection
+        /subsubsection
+        /paragraph
+        /subparagraph
+    ] def
+}}
+
 ################################
 # beginning of toplevel
 ################################
@@ -43,13 +56,9 @@
 # if encounter a src block do language identify languge and defer
 --_mtable-regex-Org=toplevel/#\+begin_src[ ]+([a-zA-Z0-9][-#+a-zA-Z0-9]*)//{tenter=srcblock}{_guest=\1,0end,}
 # sections stuff
---_mtable-regex-Org=toplevel/\*\*\*\*\*\*\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/G/
---_mtable-regex-Org=toplevel/\*\*\*\*\*\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/P/
---_mtable-regex-Org=toplevel/\*\*\*\*\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/b/
---_mtable-regex-Org=toplevel/\*\*\*\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/u/
---_mtable-regex-Org=toplevel/\*\*\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/s/
---_mtable-regex-Org=toplevel/\*\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/c/
---_mtable-regex-Org=toplevel/\*[ \t]+([[:graph:][:blank:]]+)([\n])?/\1/p/
+--_mtable-regex-Org=toplevel/(\*{1,7})[ \t]+([[:graph:][:blank:]]+)([\n])?//{{
+    \2 kindTable \1 length 1 sub get @2 _tag _commit pop
+}}
 # labels
 --_mtable-regex-Org=toplevel/<<([^>]+)>>/\1/d/
 --_mtable-regex-Org=toplevel/#\+(NAME|name):[[:blank:]]+([[:graph:][:blank:]]+)([\n])?/\2/d/