--- a/optlib/org.ctags
+++ b/optlib/org.ctags
@@ -55,13 +55,14 @@
 --_mtable-regex-Org=toplevel/[^#*<]+//
 # if encounter a src block do language identify languge and defer
 --_mtable-regex-Org=toplevel/#\+begin_src[ ]+([a-zA-Z0-9][-#+a-zA-Z0-9]*)//{tenter=srcblock}{_guest=\1,0end,}
+# labels
+--_mtable-regex-Org=toplevel/#\+(NAME|name):[[:blank:]]+([[:graph:][:blank:]]+)([\n])?/\2/d/
 # sections stuff
 --_mtable-regex-Org=toplevel/(\*{1,7})[ \t]+([[:graph:][:blank:]]+)([\n])?//{{
     \2 kindTable \1 length 1 sub get @2 _tag _commit pop
 }}
 # labels
 --_mtable-regex-Org=toplevel/<<([^>]+)>>/\1/d/
---_mtable-regex-Org=toplevel/#\+(NAME|name):[[:blank:]]+([[:graph:][:blank:]]+)([\n])?/\2/d/
 ## else do nothing (end of toplevel)
 --_mtable-regex-Org=toplevel/.//