@masatake 2022-12-27
I found the Org parser extracts many noises. It may be rewritten more line-oriented way.